\name{censor_surv}
\alias{censor_surv}
\title{censor survival event}
\usage{
censor_surv(eventtime,censortime, x)


}

\arguments{
  \item{eventtime}{vector of event times }
  \item{censortime}{vector of censor time}
  \item{x}{covariates matrix}
}


\description{
censor event time
}

