\name{column_freq}
\alias{column_freq}
\title{Calculate the Combination Frequency of Several Columns within a Data Frame}
\usage{
column_freq(x,namelist,keepID=FALSE)


}

\arguments{

  \item{x}{Input data as a data frame.}
  \item{namelist}{A list of column names that need to be counted.}
  \item{keepID}{If TRUE, the output will keep the unique ID for the column combination.}
}


\description{
Calculate the combination frequency of several columns in a data frame.
}

