\name{weibullsim_betapw}
\alias{weibullsim_betapw}
\title{Simulate Event Time with weibull Hazard Function and Piecewise Hazard Ratios}
\usage{
weibullsim_betapw(lambda,gamma,t,x,betas)


}

\arguments{



  \item{lambda}{lambda for event hazard function}
  \item{gamma}{gamma for event hazard function}
  \item{x}{design matrix}
  \item{betas}{covariates for design matrix}
  \item{t}{TIme windows for piecewise hazard ratios}
}


\description{
Simulate event time with weibull hazard function for piecewise hazard ratios
}
